refactor: break down Contact section into focused subcomponents
- Reduce Contact.tsx from 184 to 21 lines (88% reduction)
- Extract ContactInfo component for email, phone, social links, and location
- Extract ContactForm component with Formspree integration and reCAPTCHA
- Extract LocationMap component with MapLibre GL and animated marker
- Improve maintainability and testability of contact functionality
- Each component handles a single aspect of the contact section
